requirements.txt
torch
transformers
Pillow
streamlit
google-generativeai
python-dotenv
